+++ app/v1/modules/kubernetes_services/__init__.py
@@ -7,6 +7,7 @@ from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
 from .script_read_env_from import ScriptEnvironReaderFromService
 from .script_read_ingress_rules import ScriptIngressRulesReaderService
+from .script_read_ingress_tls import ScriptIngressTlsReaderService
 from .script_read_volumes_mounts import ScriptVolumeMountsReaderService
 
 
@@ -17,5 +18,6 @@ __all__ = [
     "ScriptEnvironReaderFromService",
     "ScriptEnvironReaderService",
     "ScriptIngressRulesReaderService",
+    "ScriptIngressTlsReaderService",
     "ScriptVolumeMountsReaderService",
 ]
